A new implementation of cyclic buffers
The new implementation uses a buffer twice as long as necessary and
stores each element twice. The upshot is that when reading from the
buffer, there is no need to use the mod function. There is also no
need to copy the buffer.

The new implementation doesn't replace the old implementation. It's
easy to choose between them by choosing what function to create them with.