Use elements to write pull vectors to memory.
Using parallel to write a multidimensional pull vector to memory
requires the use of the fromIndex function which generates div
and mod operations to construct a multidimensional index from a
linear index. Hence it is better to use nested loops and toIndex
which only multiplies and adds.

It is also better to use elements than M () to do the writes
since elements preserves the parallelism of the pull vector.
We essentially promise the downstream passes that the writes
are indepedent, a property that holds for these functions.