Improve layout of 'Let' when drawing ASTs.
In the AST, let expressions are represented by the 'Let' operator
which takes the RHS and a lambda expression as arguments. Hence
'let x = rhs in e' is represented as 'Let rhs (\x.e)'. This form
is harder to read when inspecting intermediate code and it also
leads to excessive indentation when the body is another let
expression. Hence we instead format a let nest of the form
'Let e1 (\x1. Let e2 (\x2. ... Let en (\xn.e)...))' as
'Let
  x1 = e1
  x2 = e2
  ...
  xn = en
In
  e'
to make it more readable.