Add pretty print for value info.
When ValueInfo is included in a pretty print it is typically adjacent
to the corresponding type. Hence it is only the ranges that are
interesting, not the ValueInfo constructors. Also, the VIProd
constructor is used for a number of purposes, including arrays and
tuples, where we can print in a more idiomatic way that is also
closer to the printing of value info in the typed AST.