Add size info to the typed abstract syntax.
This patch adds size (range) information to the typed variant
of the abstraxt syntax by redifining the Info type in Representation.
The size information is initialized to top when reifying the
program; size propagation will be the task of a separate pass.

The changes introduced in Representation leads to further
changes, mainly to strengthen contexts to enable the use
of operations from the Lattice and Eq classes on expressions
now also containing size information.

Equality is defined to require the size information to match. This
might lead to hash collisions in the CSE, but it is unlikely as the
size information is initialized to top for the corresponding type,
and the types are used when computing the hash values.