Eliminate binding time violation.
Library code (like the code in SimpleVector) executes at
specialization time (when the target version of the
Feldspar program is produced) rather than at run-time
(when the target version executes). For this reason,
library code is not allowed to inspect run-time values.
The old CSE tolerated that, but the incremental CSE
requires this rule to be obeyed.

One way of inspecting the code being generated is by
checking for equality of 'Data T' values. It would be
good to eliminate equality on 'Data T' values, but it
is needed to allow them to be in the Num class, so
we instead let equality checks raise a run time error.