Add nested tuples
This commit adds nested tuples to Feldspar. This allows Feldspar programs
to use tuples of arbitrary length, though not through the use of Haskell
tuple syntax. It is also possible to write for instance length independent
zip functions. Nested tuples are translated to the existing untyped tuples
which are already arbitrary length.

The existing support for tuples of length up to 15 is left as-is, so
using tuple syntax (,,,) yields the old tuples.