Import the sources of feldspar-compiler
This is a straightforward:

cp -a ../feldspar-compiler/{src,clib,test} .

coupled with some whitespace and hlint
fixes on top of that to bring the sources
from feldspar-compiler into feldspar-language.
The reason I squashed the whitespace and hlint
fixes in this commit is they are unlikely to break
anything so having additional whitespace commits
on top of this will just make it more tedious
to figure out that the relevant commit history
for an issue requires looking in feldspar-compiler.

The benchmarks in feldspar-compiler are not
imported in this commit since they depend
on the interaction with a custom cabal setup.
The changes in Setup.hs are imported but the
benchmarks remain in feldspar-compiler for now.