Inline Interpretation.hs into FromCore.hs
The file only contains functions used by FromCore
so inline it there for now.

This change made GHC detect that the function
"initialize" is unused. I put it on the export
list of FromCore to silence the warning for now,
but the fact that we don't initialize variables
the regular way anywhere in the compiler
is a bit strange.