travis: use cabal v2 builds
The sandbox command is removed from Cabal 3 so
switch to using "v2 builds" that have been in
Cabal since 1.24. This seems to speed up the
travis builds, the GHC 8.0.2 build took 8+ minutes
before and 5 minutes after.

More recent versions of Cabal use the v2 prefix
for the builds, but they were introduced as
"new" in Cabal 1.24, so we stick with that
for now.