Move frontend to Compiler.hs
The function frontend is really a driver
for large parts of the compiler. Move it
to Compiler.hs where the second part
of the driver, backend, resides so we get
all of it in the same place.

I tried moving backend to Frontend.hs
but that resulted in cyclic imports, so
this is the way things fit with our
current structure.