travis: build 2 files in parallel
There is support for passing -j to ghc --make
so pass -j2.

I measured on my local machine and found a slight
speedup when building from a clean state with j3
compared to j2, j4 is the same as j2 but uses
four times as much sys time. This did not translate
to Travis where the builds became much slower
than without -j so pass -j2 to get some speedup.