Remove builtin machine library
The idea was to have these functions in the program format
and be able to optimize them along with the rest of
the program. That is still a goal, but this particular
code has been in the tree for 6 years without being enabled
so remove it.

The new approach is to write these functions in C and use
External Program to splice them in in the regular
compilation pipeline.