tests: output into tmp2 directory
This restores the execution time for running
'stack test' with VSCode running to the same
time as running 'stack test' without it.

The problem turned out to be the MS C/C++ IntelliSense
plugin and file watching. Our tests output C files in
the test directory at a rapid pace which causes
IntelliSense to react to the new files and use the CPU
for non-test related activities.

The best case scenario would be to output our tests into
a temporary directory somewhere but this is the minimal
fix that I can get sane performance with. I still have
this in my settings.json:

    "files.watcherExclude": {
        "**/.git": true,
        "**/.DS_Store": true,
        "**/tmp" : true,
        "**/tmp2" : true,
        "**/.stack-work" : true,
    },
    "files.exclude": {
        "**/.git": true,
        "tmp" : true,
        "tmp2" : true,
        ".stack-work" : true,
    },