Move ONNX related files to separate directory.
The changes in this commit are necessary for Cabal to work correctly
when onnxToFeld.hs starts using parts of that library. The library
is located under 'src' and the ONNX related files were previously in
sub directories of 'src'. With that structure the onnxToFeld
executable cannot have feldspar-language as a build depend in the
Cabal file which would force the entry for that executable to include
all of the files and dependencies of the library.